When I apply this benchmark for a mali gpu g78ae, I get the result like FP32 compute 0.051 TFLOPs/s
FP16 compute 0.102 TFLOPs/s
However, what I can get from datasheet is that it could have 100GFLOPS for FP32. Then we have two ways to make the benchmark result right. One is add more fma (like 16) in the for loop and multiply it when do the math. The other is unroll the loop with this attribute.
Both will help reach the performance limit.
FP32 compute 0.102 TFLOPs/s
FP16 compute 0.204 TFLOPs/s